+++ src/Resources/skeleton/crud/test/Test.tpl.php
@@ -14,7 +14,7 @@ class <?= $class_name ?> extends WebTestCase<?= "\n" ?>
     protected function setUp(): void
     {
         $this->client = static::createClient();
-        $this->repository = (static::getContainer()->get('doctrine'))->getRepository(<?= $entity_class_name; ?>::class);
+        $this->repository = static::getContainer()->get('doctrine')->getRepository(<?= $entity_class_name; ?>::class);
 
         foreach ($this->repository->findAll() as $object) {
             $this->repository->remove($object, true);