+++ src/Resources/skeleton/twig/Component.tpl.php
@@ -1,0 +1,10 @@
+<?= "<?php\n" ?>
+
+namespace <?= $namespace; ?>;
+
+use Symfony\UX\TwigComponent\Attribute\AsTwigComponent;
+
+#[AsTwigComponent('<?= $short_name; ?>')]
+final class <?= $class_name."\n" ?>
+{
+}