+++ src/Resources/skeleton/twig/component_template.tpl.php
@@ -1,0 +1,3 @@
+<div{{ attributes }}>
+    <!-- component html -->
+</div>